// src/types/todo.ts
export interface Todo {
  id: string | number;
  title: string;
  completed: boolean;
  createdAt?: string;
  created_at?: string;
  userId?: string;
  user_id?: string;
  userEmail?: string;
  user_email?: string;
  ownerId?: string;
  ownerEmail?: string;
  owner_id?: string;
  owner_email?: string;
}